laravel_lesson 29,30
・パスワードの再設定機能を実装
・mailを送ったりするのは仮想環境では確かめにくいので保留
・ログインしないと投稿、編集、削除のボタンが表示されない

+++ resources/views/articles/show.blade.php
@@ -14,11 +14,12 @@
 
   <br/>
 
-  {!! link_to(action('ArticlesController@edit', [$article->id]), 'Edit', ['class' => 'btn btn-primary']) !!}
+  @if(Auth::check())
+    {!! link_to(action('ArticlesController@edit', [$article->id]), 'Edit', ['class' => 'btn btn-primary']) !!}
 
-  <br/>
-  <br/>
+    <br/>
+    <br/>
 
-  {!! delete_form(['articles', $article->id])  !!}
-
+    {!! delete_form(['articles', $article->id])  !!}
+  @endif
 @endsection
